requirements.txt
streamlit
google-api-python-client
google-auth
pillow
io
